feat(movie-details): add user score
