+++ packages/web/components/tmdb-card/tmdb-card.styles.tsx
@@ -1,10 +1,6 @@
 import styled from 'styled-components';
 
-interface Props {
-  vote: number;
-}
-
-export const TMDBCardStyles = styled.div<Props>`
+export const TMDBCardStyles = styled.div`
   flex-shrink: 0;
   position: relative;
   width: 220px;
@@ -73,38 +69,8 @@ export const TMDBCardStyles = styled.div<Props>`
   }
 
   .vote--container {
-    color: #fff;
-    font-size: 0.65em;
     position: absolute;
     top: 310px;
     left: 14px;
-
-    .vote {
-      border: 2px solid ${({ theme }) => theme.colors.navbarBackground};
-      background: ${(props) => (props.vote >= 70 ? '#21d07a' : '#d2d531')};
-      border-radius: 50%;
-      display: flex;
-      align-items: center;
-      justify-content: center;
-      position: relative;
-      height: 38px;
-      width: 38px;
-
-      &::before {
-        content: ' ';
-        background: ${({ theme }) => theme.colors.navbarBackground};
-        border-radius: 50%;
-        top: 0;
-        left: 0;
-        height: calc(100% - 4px);
-        width: calc(100% - 4px);
-      }
-    }
-
-    .percent {
-      position: absolute;
-      top: 14px;
-      left: 9px;
-    }
   }
 `;
