+++ packages/web/utils/get-cached-image-url.ts
@@ -1,5 +1,7 @@
 import { apiURL } from './api-url';
 
-export function getImageURL(url: string) {
-  return `${apiURL}/image-cache?i=${url}`;
+export function getImageURL(url: string | null) {
+  return url
+    ? `${apiURL}/image-cache?i=${url}`
+    : 'data:image/gif;base64,R0lGODlhAQABAIAAAP///wAAACH5BAEAAAAALAAAAAABAAEAAAICRAEAOw==';
 }
